Get file path to raw data

The raw data from the survey is stored in individual files for each survey language. These are not cleaned or harmonised, as there are small inconsistencies in coding between the languages.

Usage

gbhs_path_data(path = NULL, type = "clean", destination = NULL, ...)

Source

This function is adapted from readxl::readxl_example().

Arguments

path

Name of file in quotes with extension. If NULL, the example files will be listed.

type

type of data to look up. Either "clean" (default) or "raw"

destination

optional string indicating where to copy the file to

...

other arguments to list.files

Value

string of file path

Examples

gbhs_path_data()
#>  [1] "114338_en.tsv"  "116413_hu.tsv"  "117805_no.tsv"  "118017_fr.tsv" 
#>  [5] "119016_se.tsv"  "119157_cat.tsv" "119161_sp.tsv"  "119162_ge.tsv" 
#>  [9] "119773_dk.tsv"  "121893_ukr.tsv" "123887_it.tsv"  "125552_du.tsv" 
#> [13] "131674_ch.tsv"  "132336_tu.tsv"  "134157_en.tsv"  "135043_du.tsv" 
gbhs_path_data("114338_en.tsv")
#> [1] "/home/runner/work/_temp/Library/gbhs/exdata/clean/114338_en.tsv"
head(read.delim(gbhs_path_data("114338_en.tsv")))
#>   form_id submission_id language suffix Consent_Form Q01_Think_Brain_Frequency
#> 1  114338       4903486  english     en          Yes              Occasionally
#> 2  114338       4913285  english     en          Yes                Frequently
#> 3  114338       4933808  english     en          Yes                Frequently
#> 4  114338       4936886  english     en          Yes                Frequently
#> 5  114338       4948775  english     en          Yes                    Rarely
#> 6  114338       4948924  english     en          Yes                Frequently
#>   Q02_DietInfl Q02_EducationInfl Q02_GeneticsInfl Q02_IncomeInfl
#> 1     Moderate            Strong         Moderate           Weak
#> 2  Very strong       Very strong      Very strong    Very strong
#> 3  Very strong            Strong           Strong       Moderate
#> 4  Very strong          Moderate           Strong           Weak
#> 5       Strong       Very strong           Strong       Moderate
#> 6       Strong      No influence           Strong   No influence
#>   Q02_LifeGoalsInfl Q02_PhyEnvironmentInfl Q02_PhyHealthInfl Q02_ProfessionInfl
#> 1       Very strong                 Strong            Strong           Moderate
#> 2       Very strong            Very strong       Very strong        Very strong
#> 3          Moderate                 Strong       Very strong             Strong
#> 4            Strong            Very strong       Very strong           Moderate
#> 5            Strong                 Strong            Strong        Very strong
#> 6          Moderate               Moderate       Very strong       No influence
#>   Q02_SleepHabitsInfl Q02_SocEnvironmentInfl Q02_SubstanceInfl Q03_Adolescence
#> 1              Strong                 Strong          Moderate  Very important
#> 2         Very strong            Very strong       Very strong  Very important
#> 3         Very strong                 Strong            Strong  Very important
#> 4         Very strong            Very strong       Very strong  Very important
#> 5              Strong                 Strong            Strong  Very important
#> 6         Very strong               Moderate            Strong       Important
#>        Q03_Child  Q03_MiddleAge   Q03_OlderAge       Q03_Womb Q03_YoungAdult
#> 1 Very important Very important Very important      Important Very important
#> 2 Very important Very important Very important      Important Very important
#> 3 Very important      Important Very important      Important Very important
#> 4 Very important Very important Very important Very important Very important
#> 5 Very important      Important Very important Very important      Important
#> 6      Important Very important Very important Very important      Important
#>                                                                                                               q04_brain_disease
#> 1                                                           Addiction;Alzheimer;Anxiety;Bipolar;Cancer;Depression;Schizophrenia
#> 2 Addiction;Alzheimer;Anxiety;Arthritis;Bipolar;Cancer;Depression;Diabetes;Hypertension;Migraine;Parkinson;Schizophrenia;Stroke
#> 3                                        Addiction;Alzheimer;Anxiety;Bipolar;Depression;Migraine;Parkinson;Schizophrenia;Stroke
#> 4                                               Addiction;Alzheimer;Anxiety;Bipolar;Depression;Migraine;Parkinson;Schizophrenia
#> 5                                        Addiction;Alzheimer;Anxiety;Bipolar;Depression;Migraine;Parkinson;Schizophrenia;Stroke
#> 6                  Addiction;Alzheimer;Anxiety;Bipolar;Depression;Diabetes;Hypertension;Migraine;Parkinson;Schizophrenia;Stroke
#>   Q05_AlcoholGen Q05_ExcerciseGen Q05_HealthyDietGen Q05_IntellectGen
#> 1   Occasionally       Frequently         Frequently     Occasionally
#> 2         Rarely       Frequently         Frequently       Frequently
#> 3   Occasionally     Occasionally         Frequently           Rarely
#> 4         Rarely       Frequently         Frequently     Occasionally
#> 5   Occasionally       Frequently         Frequently       Frequently
#> 6         Rarely     Occasionally       Occasionally     Occasionally
#>   Q05_ProtectionGen Q05_RelaxGen Q05_SleepGen Q05_SmokeGen Q05_SocialiseGen
#> 1        Frequently   Frequently   Frequently       Rarely     Occasionally
#> 2             Never       Rarely Occasionally        Never       Frequently
#> 3      Occasionally       Rarely   Frequently Occasionally       Frequently
#> 4        Frequently   Frequently   Frequently        Never       Frequently
#> 5        Frequently       Rarely Occasionally        Never     Occasionally
#> 6        Frequently       Rarely   Frequently        Never     Occasionally
#>   Q05_SupplementsGen Q05_WorkLifeBalanceGen Q06_ExcercisePurp
#> 1       Occasionally           Occasionally              <NA>
#> 2              Never             Frequently              <NA>
#> 3             Rarely                 Rarely      Occasionally
#> 4         Frequently             Frequently        Frequently
#> 5       Occasionally           Occasionally             Never
#> 6             Rarely             Frequently      Occasionally
#>   Q06_HealthyDietPurp Q06_OtherPurp Q06_ProtectionPurp Q06_RelaxPurp
#> 1                <NA>          <NA>               <NA>          <NA>
#> 2                <NA>          <NA>               <NA>          <NA>
#> 3        Occasionally          <NA>       Occasionally        Rarely
#> 4          Frequently    Frequently         Frequently    Frequently
#> 5               Never          <NA>         Frequently    Frequently
#> 6        Occasionally          <NA>         Frequently        Rarely
#>   Q06_SleepPurp Q06_SocialisePurp Q06_SupplementsPurp Q06_WorkLifeBalancePurp
#> 1          <NA>              <NA>                <NA>                    <NA>
#> 2          <NA>              <NA>                <NA>                    <NA>
#> 3  Occasionally             Never              Rarely                   Never
#> 4    Frequently      Occasionally        Occasionally                  Rarely
#> 5    Frequently            Rarely        Occasionally            Occasionally
#> 6    Frequently        Frequently        Occasionally              Frequently
#>                        Q07_Test_Criteria Q08_Take_Test     Q09a_Test_Reason
#> 1                  Accurate;Cheap;NoPain  Probably not                 <NA>
#> 2                   Accurate;Cheap;Quick    Definitely Learn about the risk
#> 3                         Accurate;Cheap    Definitely Learn about the risk
#> 4   NoPain;OnlineCriteria;SocialSecurity      Probably  Information;Respond
#> 5 Accurate;OnlineCriteria;SocialSecurity  Probably not                 <NA>
#> 6          Accurate;Cheap;SocialSecurity    Definitely          Information
#>   Q09b_Test_Nonpreventable       Q09c_No_Test Q10_FamilyFriends Q10_Library
#> 1                     <NA>      Other reasons            Likely  Definitely
#> 2                 Probably               <NA>            Likely      Likely
#> 3                 Probably               <NA>        Definitely  Definitely
#> 4                 Probably               <NA>        Definitely  Definitely
#> 5                     <NA> Don't want to know    Definitely not  Definitely
#> 6               Definitely               <NA>        Definitely  Definitely
#>   Q10_LifestyleChange   Q10_Plan Q10_ProfessHelp Q11_BrainStimuli
#> 1          Definitely Definitely      Definitely      Very likely
#> 2              Likely     Likely          Likely      Very likely
#> 3          Definitely Definitely      Definitely           Likely
#> 4          Definitely Definitely      Definitely      Very likely
#> 5          Definitely Definitely      Definitely       Already do
#> 6          Definitely   Unlikely      Definitely           Likely
#>   Q11_EatHealthy Q11_ExcerciseMore Q11_ImprSleepingHabits Q11_MoreCulture
#> 1    Very likely       Very likely            Very likely     Very likely
#> 2    Very likely       Very likely            Very likely     Very likely
#> 3    Very likely       Very likely            Very likely          Likely
#> 4    Very likely       Very likely            Very likely     Very likely
#> 5    Very likely        Already do            Very likely      Already do
#> 6         Likely       Very likely            Very likely     Very likely
#>   Q11_MoreRelax Q11_MoreSocial Q11_NoAlcohol Q11_NoSmoking Q11_OtherChanges
#> 1   Very likely         Likely   Very likely   Very likely      Very likely
#> 2   Very likely    Very likely   Very likely   Very likely             <NA>
#> 3        Likely         Likely        Likely        Likely             <NA>
#> 4   Very likely    Very likely   Very likely   Very likely             <NA>
#> 5   Very likely         Likely   Very likely    Already do             <NA>
#> 6   Very likely         Likely        Likely    Already do             <NA>
#>           Q12_MotivateChange_Category             Q13_NoChange_Motivation
#> 1  Diagnose;FamilySupport;WorseMemory      GiveupHobby;NotHelp;StartNoFun
#> 2 Diagnose;PersonalAdvice;WorseMemory              Alone;Expensive;NoTime
#> 3    Diagnose;ImpactKnown;WorseMemory             Expensive;NoInfo;NoNeed
#> 4       Diagnose;Family;FamilySupport                    Expensive;NoInfo
#> 5    Diagnose;ImpactKnown;WorseMemory           GiveupHobby;NoInfo;NoTime
#> 6 Diagnose;PersonalAdvice;WorseMemory GiveupHobby;NoMotivation;StartNoFun
#>                 Q14_ChangeAssistance Q15_FreetestAuth Q15_InformAuth
#> 1      GP;GroupSupport;SupportFamily               No            Yes
#> 2 Checkup;GP;MobileApp;SupportFamily              Yes            Yes
#> 3                         Checkup;GP              Yes            Yes
#> 4           Checkup;GP;SupportFamily              Yes            Yes
#> 5           Checkup;GP;SupportFamily              Yes            Yes
#> 6            Checkup;GP;GroupSupport              Yes            Yes
#>   Q15_InsAuth Q15_RelaxAuth Q15_SubsAuth Q15_TaxAuth Q16_BrainHealthSpecialist
#> 1         Yes           Yes          Yes         Yes              Great extent
#> 2         Yes           Yes          Yes         Yes              Great extent
#> 3         Yes           Yes          Yes         Yes              Great extent
#> 4         Yes           Yes          Yes         Yes              Great extent
#> 5         Yes           Yes          Yes     Unknown              Great extent
#> 6         Yes           Yes          Yes          No              Great extent
#>         Q16_GP Q16_HealthAuthWebsites Q16_Newspaper Q16_OnlineForum
#> 1 Great extent           Great extent    Moderately      Moderately
#> 2 Great extent           Great extent    Not at all      Moderately
#> 3   Moderately           Great extent    Moderately      Not at all
#> 4   Not at all           Great extent    Moderately      Not at all
#> 5 Great extent           Great extent    Moderately      Moderately
#> 6 Great extent           Great extent    Not at all      Moderately
#>   Q16_PublicResearch Q16_ScientificJournals Q16_SocialMedia Q16_TVRadio
#> 1         Moderately           Great extent      Moderately  Moderately
#> 2       Great extent           Great extent      Moderately  Moderately
#> 3       Great extent           Great extent      Not at all  Moderately
#> 4         Moderately           Great extent      Not at all  Moderately
#> 5       Great extent           Great extent      Not at all  Moderately
#> 6       Great extent           Great extent      Not at all  Moderately
#>   Q17_BrainResearch_Part Q18_Age Q19_Residence Q20_Sex Q21_Education_Coded
#> 1                     No   26-40        Norway    Male          GradSchool
#> 2                    Yes   26-40         Spain    Male          GradSchool
#> 3                     No   26-40   Switzerland    Male          GradSchool
#> 4                     No   26-40         Spain  Female          GradSchool
#> 5                     No   41-60        Norway  Female          GradSchool
#> 6                    Yes   41-60        Norway    Male          GradSchool
#>   Q22_Relationship_Status        Q23_Vocation_categories Q24_Exp_HealthCare
#> 1                  Single             Student;Unemployed                Yes
#> 2               Unmarried                       Employed                Yes
#> 3         DomesticPartner Disabled;Employed;Stay-at-home                Yes
#> 4         DomesticPartner                       Employed                 No
#> 5         DomesticPartner                       Employed                 No
#> 6               Unmarried                       Employed                 No
#>   Q25_Cognitive_Health Q26_Emotional_Health Q27_Illness Q28_BrainDisease_Care
#> 1            Excellent            Excellent          No                    No
#> 2                 <NA>                 <NA>          No                    No
#> 3                 <NA>                 <NA>         Yes                    No
#> 4                 <NA>                 <NA>         Yes                    No
#> 5            Excellent                 <NA>          No                    No
#> 6            Excellent                 <NA>         Yes                    No
#>   Q02_DietInfl_dummy Q02_EducationInfl_dummy Q02_GeneticsInfl_dummy
#> 1                  3                       2                      3
#> 2                  1                       1                      1
#> 3                  1                       2                      2
#> 4                  1                       3                      2
#> 5                  2                       1                      2
#> 6                  2                       5                      2
#>   Q02_IncomeInfl_dummy Q02_LifeGoalsInfl_dummy Q02_PhyEnvironmentInfl_dummy
#> 1                    4                       1                            2
#> 2                    1                       1                            1
#> 3                    3                       3                            2
#> 4                    4                       2                            1
#> 5                    3                       2                            2
#> 6                    5                       3                            3
#>   Q02_PhyHealthInfl_dummy Q02_ProfessionInfl_dummy Q02_SleepHabitsInfl_dummy
#> 1                       2                        3                         2
#> 2                       1                        1                         1
#> 3                       1                        2                         1
#> 4                       1                        3                         1
#> 5                       2                        1                         2
#> 6                       1                        5                         1
#>   Q02_SocEnvironmentInfl_dummy Q02_SubstanceInfl_dummy Q03_Adolescence_dummy
#> 1                            2                       3                     1
#> 2                            1                       1                     1
#> 3                            2                       2                     1
#> 4                            1                       1                     1
#> 5                            2                       2                     1
#> 6                            3                       2                     2
#>   Q03_Child_dummy Q03_MiddleAge_dummy Q03_OlderAge_dummy Q03_Womb_dummy
#> 1               1                   1                  1              2
#> 2               1                   1                  1              2
#> 3               1                   2                  1              2
#> 4               1                   1                  1              1
#> 5               1                   2                  1              1
#> 6               2                   1                  1              1
#>   Q03_YoungAdult_dummy Q05_AlcoholGen_dummy Q05_ExcerciseGen_dummy
#> 1                    1                    2                      1
#> 2                    1                    3                      1
#> 3                    1                    2                      2
#> 4                    1                    3                      1
#> 5                    2                    2                      1
#> 6                    2                    3                      2
#>   Q05_HealthyDietGen_dummy Q05_IntellectGen_dummy Q05_ProtectionGen_dummy
#> 1                        1                      2                       1
#> 2                        1                      1                       4
#> 3                        1                      3                       2
#> 4                        1                      2                       1
#> 5                        1                      1                       1
#> 6                        2                      2                       1
#>   Q05_RelaxGen_dummy Q05_SleepGen_dummy Q05_SmokeGen_dummy
#> 1                  1                  1                  3
#> 2                  3                  2                  4
#> 3                  3                  1                  2
#> 4                  1                  1                  4
#> 5                  3                  2                  4
#> 6                  3                  1                  4
#>   Q05_SocialiseGen_dummy Q05_SupplementsGen_dummy Q05_WorkLifeBalanceGen_dummy
#> 1                      2                        2                            2
#> 2                      1                        4                            1
#> 3                      1                        3                            3
#> 4                      1                        1                            1
#> 5                      2                        2                            2
#> 6                      2                        3                            1
#>   Q06_ExcercisePurp_dummy Q06_HealthyDietPurp_dummy Q06_OtherPurp_dummy
#> 1                      NA                        NA                  NA
#> 2                      NA                        NA                  NA
#> 3                       2                         2                  NA
#> 4                       1                         1                   1
#> 5                       4                         4                  NA
#> 6                       2                         2                  NA
#>   Q06_ProtectionPurp_dummy Q06_RelaxPurp_dummy Q06_SleepPurp_dummy
#> 1                       NA                  NA                  NA
#> 2                       NA                  NA                  NA
#> 3                        2                   3                   2
#> 4                        1                   1                   1
#> 5                        1                   1                   1
#> 6                        1                   3                   1
#>   Q06_SocialisePurp_dummy Q06_SupplementsPurp_dummy
#> 1                      NA                        NA
#> 2                      NA                        NA
#> 3                       4                         3
#> 4                       2                         2
#> 5                       3                         2
#> 6                       1                         2
#>   Q06_WorkLifeBalancePurp_dummy Q10_FamilyFriends_dummy Q10_Library_dummy
#> 1                            NA                       2                 1
#> 2                            NA                       2                 2
#> 3                             4                       1                 1
#> 4                             3                       1                 1
#> 5                             2                       4                 1
#> 6                             1                       1                 1
#>   Q10_LifestyleChange_dummy Q10_Plan_dummy Q10_ProfessHelp_dummy
#> 1                         1              1                     1
#> 2                         2              2                     2
#> 3                         1              1                     1
#> 4                         1              1                     1
#> 5                         1              1                     1
#> 6                         1              3                     1
#>   Q11_BrainStimuli_dummy Q11_EatHealthy_dummy Q11_ExcerciseMore_dummy
#> 1                      1                    1                       1
#> 2                      1                    1                       1
#> 3                      2                    1                       1
#> 4                      1                    1                       1
#> 5                      3                    1                       3
#> 6                      2                    2                       1
#>   Q11_ImprSleepingHabits_dummy Q11_MoreCulture_dummy Q11_MoreRelax_dummy
#> 1                            1                     1                   1
#> 2                            1                     1                   1
#> 3                            1                     2                   2
#> 4                            1                     1                   1
#> 5                            1                     3                   1
#> 6                            1                     1                   1
#>   Q11_MoreSocial_dummy Q11_NoAlcohol_dummy Q11_NoSmoking_dummy
#> 1                    2                   1                   1
#> 2                    1                   1                   1
#> 3                    2                   2                   2
#> 4                    1                   1                   1
#> 5                    2                   1                   3
#> 6                    2                   2                   3
#>   Q11_OtherChanges_dummy Q15_FreetestAuth_dummy Q15_InformAuth_dummy
#> 1                      1                      2                    1
#> 2                     NA                      1                    1
#> 3                     NA                      1                    1
#> 4                     NA                      1                    1
#> 5                     NA                      1                    1
#> 6                     NA                      1                    1
#>   Q15_InsAuth_dummy Q15_RelaxAuth_dummy Q15_SubsAuth_dummy Q15_TaxAuth_dummy
#> 1                 1                   1                  1                 1
#> 2                 1                   1                  1                 1
#> 3                 1                   1                  1                 1
#> 4                 1                   1                  1                 1
#> 5                 1                   1                  1                 3
#> 6                 1                   1                  1                 2
#>   Q16_BrainHealthSpec_dummy Q16_GP_dummy Q16_HealthAuthWebsites_dummy
#> 1                         1            1                            1
#> 2                         1            1                            1
#> 3                         1            2                            1
#> 4                         1            3                            1
#> 5                         1            1                            1
#> 6                         1            1                            1
#>   Q16_Newspaper_dummy Q16_OnlineForum_dummy Q16_PublicResearch_dummy
#> 1                   2                     2                        2
#> 2                   3                     2                        1
#> 3                   2                     3                        1
#> 4                   2                     3                        2
#> 5                   2                     2                        1
#> 6                   3                     2                        1
#>   Q16_ScientificJournals_dummy Q16_SocialMedia_dummy Q16_TVRadio_dummy
#> 1                            1                     2                 2
#> 2                            1                     2                 2
#> 3                            1                     3                 2
#> 4                            1                     3                 2
#> 5                            1                     3                 2
#> 6                            1                     3                 2
#>   Q18_Age_dummy Q19_Residence_dummy Q20_Sex_dummy Q21_Education_Coded_dummy
#> 1             2                 127             2                         5
#> 2             2                 157             2                         5
#> 3             2                 161             2                         5
#> 4             2                 157             1                         5
#> 5             3                 127             1                         5
#> 6             3                 127             2                         5
#>   Q22_Relationship_Status_dummy
#> 1                             1
#> 2                             2
#> 3                             3
#> 4                             3
#> 5                             3
#> 6                             2
#>                                                                                             Q06_BrainHealthPurpose_Spec
#> 1                                                                           Spend time alone and enjoy ones own company
#> 2                                                                                                                  <NA>
#> 3                                                                                                                  <NA>
#> 4 Brain games, investigating about brain health, listnening to podcast related to brain activity and enhance my memory.
#> 5                                                                                                                  <NA>
#> 6                                                                                                                  <NA>
#>   Q07_Test_Criteria_Spec Q09a_Test_Reason_Spec
#> 1                   <NA>                  <NA>
#> 2                   <NA>                  <NA>
#> 3                   <NA>                  <NA>
#> 4                   <NA>                  <NA>
#> 5                   <NA>                  <NA>
#> 6                   <NA>                  <NA>
#>                      Q09c_OtherReason_Spec
#> 1 There is enough health hysteria as it is
#> 2                                     <NA>
#> 3                                     <NA>
#> 4                                     <NA>
#> 5                                     <NA>
#> 6                                     <NA>
#>                                       Q10_TestReaction_Spec
#> 1 Dont spend time on bad friends and meaningless activities
#> 2                                                      <NA>
#> 3                                                      <NA>
#> 4             Trust that everything is going to be alright.
#> 5                                                      <NA>
#> 6                                                      <NA>
#>       Q11_OtherChanges_Spec Q12_OtherChangeMotivation_Spec
#> 1 Spend more time in nature                           <NA>
#> 2                      <NA>                           <NA>
#> 3                      <NA>                           <NA>
#> 4                      <NA>                           <NA>
#> 5                      <NA>                           <NA>
#> 6                      <NA>                           <NA>
#>   Q13_SpecifyNoMotivation_Spec Q14_OtherAssistance_Spec
#> 1                         <NA>                     <NA>
#> 2                         <NA>                     <NA>
#> 3                         <NA>                     <NA>
#> 4                         <NA>                     <NA>
#> 5                         <NA>                     <NA>
#> 6                         <NA>                     <NA>
head(read.delim(gbhs_path_data("114338_en.tsv", "raw")))
#>   form_id submission_id Consent_Form_en Q01_Think_Brain_Frequency_en
#> 1  114338       9052580  Consent_Yes_en                 Occasionally
#> 2  114338       9051492  Consent_Yes_en                   Frequently
#> 3  114338       9050113  Consent_Yes_en                 Occasionally
#> 4  114338       9050041  Consent_Yes_en                 Occasionally
#> 5  114338       9049694  Consent_Yes_en                       Rarely
#> 6  114338       9049513  Consent_Yes_en                       Rarely
#>   Q02_DietInfl_en Q02_DietInfl_en_dummy Q02_EducationInfl_en
#> 1          Strong                     2             Moderate
#> 2      VeryStrong                     1               Strong
#> 3            Weak                     4               Strong
#> 4          Strong                     2               Strong
#> 5        Moderate                     3             Moderate
#> 6            Weak                     4             Moderate
#>   Q02_EducationInfl_en_dummy Q02_GeneticsInfl_en Q02_GeneticsInfl_en_dummy
#> 1                          3          VeryStrong                         1
#> 2                          2          VeryStrong                         1
#> 3                          2              Strong                         2
#> 4                          2              Strong                         2
#> 5                          3            Moderate                         3
#> 6                          3              Strong                         2
#>   Q02_IncomeInfl_en Q02_IncomeInfl_en_dummy Q02_LifeGoalsInfl_en
#> 1              Weak                       4               Strong
#> 2            Strong                       2               Strong
#> 3          Moderate                       3               Strong
#> 4          Moderate                       3             Moderate
#> 5          Moderate                       3               Strong
#> 6            Strong                       2             Moderate
#>   Q02_LifeGoalsInfl_en_dummy Q02_PhyEnvironmentInfl_en
#> 1                          2                  Moderate
#> 2                          2                VeryStrong
#> 3                          2                VeryStrong
#> 4                          3                    Strong
#> 5                          2                  Moderate
#> 6                          3                    Strong
#>   Q02_PhyEnvironmentInfl_en_dummy Q02_PhyHealthInfl_en
#> 1                               3           VeryStrong
#> 2                               1           VeryStrong
#> 3                               1               Strong
#> 4                               2               Strong
#> 5                               3                 <NA>
#> 6                               2             Moderate
#>   Q02_PhyHealthInfl_en_dummy Q02_ProfessionInfl_en Q02_ProfessionInfl_en_dummy
#> 1                          1              Moderate                           3
#> 2                          1              Moderate                           3
#> 3                          2              Moderate                           3
#> 4                          2              Moderate                           3
#> 5                         NA              Moderate                           3
#> 6                          3              Moderate                           3
#>   Q02_SleepHabitsInfl_en Q02_SleepHabitsInfl_en_dummy Q02_SocEnvironmentInfl_en
#> 1             VeryStrong                            1                  Moderate
#> 2             VeryStrong                            1                VeryStrong
#> 3             VeryStrong                            1                    Strong
#> 4                 Strong                            2                  Moderate
#> 5               Moderate                            3                    Strong
#> 6               Moderate                            3                  Moderate
#>   Q02_SocEnvironmentInfl_en_dummy Q02_SubstanceInfl_en
#> 1                               3           VeryStrong
#> 2                               1           VeryStrong
#> 3                               2               Strong
#> 4                               3           VeryStrong
#> 5                               2             Moderate
#> 6                               3               Strong
#>   Q02_SubstanceInfl_en_dummy Q03_Adolescence_en Q03_Adolescence_en_dummy
#> 1                          1          Important                        2
#> 2                          1          Important                        2
#> 3                          2       NotImportant                        4
#> 4                          1      VeryImportant                        1
#> 5                          3      VeryImportant                        1
#> 6                          2          Important                        2
#>    Q03_Child_en Q03_Child_en_dummy Q03_MiddleAge_en Q03_MiddleAge_en_dummy
#> 1     Important                  2     ModImportant                      3
#> 2 VeryImportant                  1    VeryImportant                      1
#> 3  ModImportant                  3        Important                      2
#> 4 VeryImportant                  1        Important                      2
#> 5 VeryImportant                  1    VeryImportant                      1
#> 6 VeryImportant                  1    VeryImportant                      1
#>   Q03_OlderAge_en Q03_OlderAge_en_dummy   Q03_Womb_en Q03_Womb_en_dummy
#> 1       Important                     2  ModImportant                 3
#> 2   VeryImportant                     1 VeryImportant                 1
#> 3       Important                     2     Important                 2
#> 4       Important                     2     Important                 2
#> 5   VeryImportant                     1  ModImportant                 3
#> 6   VeryImportant                     1 VeryImportant                 1
#>   Q03_YoungAdult_en Q03_YoungAdult_en_dummy
#> 1      ModImportant                       3
#> 2     VeryImportant                       1
#> 3      ModImportant                       3
#> 4     VeryImportant                       1
#> 5     VeryImportant                       1
#> 6      ModImportant                       3
#>                                                                                  Q04_Brain_Disease_en
#> 1              Addiction;Alzheimer;Anxiety;Bipolar;Depression;Migraine;Parkinson;Schizophrenia;Stroke
#> 2                      Addiction;Alzheimer;Bipolar;Depression;Migraine;Parkinson;Schizophrenia;Stroke
#> 3       Addiction;Alzheimer;Anxiety;Bipolar;Cancer;Depression;Migraine;Parkinson;Schizophrenia;Stroke
#> 4 Addiction;Alzheimer;Anxiety;Bipolar;Depression;Hypertension;Migraine;Parkinson;Schizophrenia;Stroke
#> 5                     Addiction;Alzheimer;Anxiety;Bipolar;Depression;Migraine;Parkinson;Schizophrenia
#> 6                     Addiction;Alzheimer;Anxiety;Bipolar;Depression;Migraine;Parkinson;Schizophrenia
#>   Q05_AlcoholGen_en Q05_AlcoholGen_en_dummy Q05_ExcerciseGen_en
#> 1            Rarely                       3          Frequently
#> 2            Rarely                       3          Frequently
#> 3            Rarely                       3        Occasionally
#> 4      Occasionally                       2          Frequently
#> 5      Occasionally                       2          Frequently
#> 6        Frequently                       1          Frequently
#>   Q05_ExcerciseGen_en_dummy Q05_HealthyDietGen_en Q05_HealthyDietGen_en_dummy
#> 1                         1            Frequently                           1
#> 2                         1          Occasionally                           2
#> 3                         2            Frequently                           1
#> 4                         1            Frequently                           1
#> 5                         1          Occasionally                           2
#> 6                         1                Rarely                           3
#>   Q05_IntellectGen_en Q05_IntellectGen_en_dummy Q05_ProtectionGen_en
#> 1          Frequently                         1           Frequently
#> 2          Frequently                         1           Frequently
#> 3              Rarely                         3                Never
#> 4          Frequently                         1         Occasionally
#> 5              Rarely                         3           Frequently
#> 6              Rarely                         3           Frequently
#>   Q05_ProtectionGen_en_dummy Q05_RelaxGen_en Q05_RelaxGen_en_dummy
#> 1                          1    Occasionally                     2
#> 2                          1      Frequently                     1
#> 3                          4      Frequently                     1
#> 4                          2    Occasionally                     2
#> 5                          1    Occasionally                     2
#> 6                          1    Occasionally                     2
#>   Q05_SleepGen_en Q05_SleepGen_en_dummy Q05_SmokeGen_en Q05_SmokeGen_en_dummy
#> 1    Occasionally                     2           Never                     4
#> 2      Frequently                     1           Never                     4
#> 3    Occasionally                     2           Never                     4
#> 4    Occasionally                     2           Never                     4
#> 5    Occasionally                     2           Never                     4
#> 6      Frequently                     1           Never                     4
#>   Q05_SocialiseGen_en Q05_SocialiseGen_en_dummy Q05_SupplementsGen_en
#> 1              Rarely                         3          Occasionally
#> 2              Rarely                         3            Frequently
#> 3          Frequently                         1                 Never
#> 4        Occasionally                         2          Occasionally
#> 5        Occasionally                         2                Rarely
#> 6          Frequently                         1                Rarely
#>   Q05_SupplementsGen_en_dummy Q05_WorkLifeBalanceGen_en
#> 1                           2              Occasionally
#> 2                           1                Frequently
#> 3                           4              Occasionally
#> 4                           2              Occasionally
#> 5                           3              Occasionally
#> 6                           3              Occasionally
#>   Q05_WorkLifeBalanceGen_en_dummy Q06_BrainHealth_Purpose_Spec_en
#> 1                               2                            <NA>
#> 2                               1                            <NA>
#> 3                               2                            <NA>
#> 4                               2                            <NA>
#> 5                               2                            <NA>
#> 6                               2                            <NA>
#>   Q06_ExcercisePurp_en Q06_ExcercisePurp_en_dummy Q06_HealthyDietPurp_en
#> 1           Frequently                          1             Frequently
#> 2           Frequently                          1           Occasionally
#> 3           Frequently                          1           Occasionally
#> 4         Occasionally                          2           Occasionally
#> 5         Occasionally                          2           Occasionally
#> 6           Frequently                          1           Occasionally
#>   Q06_HealthyDietPurp_en_dummy Q06_OtherPurp_en Q06_OtherPurp_en_dummy
#> 1                            1             <NA>                     NA
#> 2                            2             <NA>                     NA
#> 3                            2             <NA>                     NA
#> 4                            2             <NA>                     NA
#> 5                            2     Occasionally                      2
#> 6                            2       Frequently                      1
#>   Q06_ProtectionPurp_en Q06_ProtectionPurp_en_dummy Q06_RelaxPurp_en
#> 1            Frequently                           1     Occasionally
#> 2            Frequently                           1     Occasionally
#> 3                 Never                           4           Rarely
#> 4          Occasionally                           2       Frequently
#> 5            Frequently                           1     Occasionally
#> 6            Frequently                           1       Frequently
#>   Q06_RelaxPurp_en_dummy Q06_SleepPurp_en Q06_SleepPurp_en_dummy
#> 1                      2     Occasionally                      2
#> 2                      2       Frequently                      1
#> 3                      3     Occasionally                      2
#> 4                      1       Frequently                      1
#> 5                      2     Occasionally                      2
#> 6                      1       Frequently                      1
#>   Q06_SocialisePurp_en Q06_SocialisePurp_en_dummy Q06_SupplementsPurp_en
#> 1         Occasionally                          2             Frequently
#> 2               Rarely                          3             Frequently
#> 3         Occasionally                          2                  Never
#> 4         Occasionally                          2                 Rarely
#> 5         Occasionally                          2           Occasionally
#> 6           Frequently                          1                 Rarely
#>   Q06_SupplementsPurp_en_dummy Q06_WorkLifeBalancePurp_en
#> 1                            1               Occasionally
#> 2                            1                 Frequently
#> 3                            4               Occasionally
#> 4                            3                 Frequently
#> 5                            2               Occasionally
#> 6                            3                 Frequently
#>   Q06_WorkLifeBalancePurp_en_dummy           Q07_Test_Criteria_en
#> 1                                2           Accurate;Cheap;Quick
#> 2                                1  Accurate;Cheap;SocialSecurity
#> 3                                2  Accurate;Cheap;SocialSecurity
#> 4                                1 Accurate;NoPain;SocialSecurity
#> 5                                2           Accurate;Cheap;Quick
#> 6                                1  Accurate;Cheap;OnlineCriteria
#>   Q07_Test_Criteria_Spec_en Q08_Take_Test_en   Q09a_Test_Reason_en
#> 1                      <NA>       Definitely               Respond
#> 2                      <NA>       Definitely   Information;Prepare
#> 3                      <NA>         Probably Information;LearnRisk
#> 4                      <NA>       Definitely     LearnRisk;Prepare
#> 5                      <NA>       Definitely   Information;Respond
#> 6                      <NA>       Definitely   Information;Respond
#>   Q09a_Test_Reason_Spec_en Q09b_Test_Nonpreventable_en Q09c_No_Test_en
#> 1                     <NA>                      DefYes            <NA>
#> 2                     <NA>                 ProbablyYes            <NA>
#> 3                     <NA>                      DefYes            <NA>
#> 4                     <NA>                 ProbablyYes            <NA>
#> 5                     <NA>                 ProbablyYes            <NA>
#> 6                     <NA>                 ProbablyYes            <NA>
#>   Q09c_OtherReason_Spec_en Q10_FamilyFriends_en Q10_FamilyFriends_en_dummy
#> 1                     <NA>               Likely                          2
#> 2                     <NA>               DefYes                          1
#> 3                     <NA>               DefYes                          1
#> 4                     <NA>             Unlikely                          3
#> 5                     <NA>             Unlikely                          3
#> 6                     <NA>               Likely                          2
#>   Q10_Library_en Q10_Library_en_dummy Q10_LifestyleChange_en
#> 1         DefYes                    1                 DefYes
#> 2         DefYes                    1                 DefYes
#> 3         Likely                    2               Unlikely
#> 4         DefYes                    1                 Likely
#> 5         DefYes                    1                 DefYes
#> 6         DefYes                    1                 DefYes
#>   Q10_LifestyleChange_en_dummy Q10_Plan_en Q10_Plan_en_dummy Q10_ProfessHelp_en
#> 1                            1    Unlikely                 3             DefYes
#> 2                            1      DefYes                 1             DefYes
#> 3                            3    Unlikely                 3           Unlikely
#> 4                            2      DefYes                 1             DefYes
#> 5                            1      DefYes                 1             DefYes
#> 6                            1      DefYes                 1             DefYes
#>   Q10_ProfessHelp_en_dummy Q10_TestReactionSpec_en Q11_BrainStimuli_en
#> 1                        1                    <NA>           AlreadyDo
#> 2                        1                     Cry          VeryLikely
#> 3                        3                    <NA>              Likely
#> 4                        1                    <NA>           AlreadyDo
#> 5                        1                    <NA>          VeryLikely
#> 6                        1                    <NA>          VeryLikely
#>   Q11_BrainStimuli_en_dummy Q11_EatHealthy_en Q11_EatHealthy_en_dummy
#> 1                         3        VeryLikely                       1
#> 2                         1        VeryLikely                       1
#> 3                         2            Likely                       2
#> 4                         3        VeryLikely                       1
#> 5                         1        VeryLikely                       1
#> 6                         1        VeryLikely                       1
#>   Q11_ExcerciseMore_en Q11_ExcerciseMore_en_dummy Q11_ImprSleepingHabits_en
#> 1           VeryLikely                          1                    Likely
#> 2           VeryLikely                          1                VeryLikely
#> 3               Likely                          2                VeryLikely
#> 4           VeryLikely                          1                VeryLikely
#> 5           VeryLikely                          1                VeryLikely
#> 6           VeryLikely                          1                VeryLikely
#>   Q11_ImprSleepingHabits_en_dummy Q11_MoreCulture_en Q11_MoreCulture_en_dummy
#> 1                               2             Likely                        2
#> 2                               1             Likely                        2
#> 3                               1             Likely                        2
#> 4                               1         VeryLikely                        1
#> 5                               1         VeryLikely                        1
#> 6                               1             Likely                        2
#>   Q11_MoreRelax_en Q11_MoreRelax_en_dummy Q11_MoreSocial_en
#> 1           Likely                      2            Likely
#> 2       VeryLikely                      1            Likely
#> 3       VeryLikely                      1            Likely
#> 4           Likely                      2            Likely
#> 5       VeryLikely                      1        VeryLikely
#> 6       VeryLikely                      1              <NA>
#>   Q11_MoreSocial_en_dummy Q11_NoAlcohol_en Q11_NoAlcohol_en_dummy
#> 1                       2       VeryLikely                      1
#> 2                       2       VeryLikely                      1
#> 3                       2         Unlikely                      4
#> 4                       2           Likely                      2
#> 5                       1       VeryLikely                      1
#> 6                      NA           Likely                      2
#>   Q11_NoSmoking_en Q11_NoSmoking_en_dummy Q11_OtherChanges_en
#> 1       VeryLikely                      1                <NA>
#> 2       VeryLikely                      1                <NA>
#> 3       VeryLikely                      1                <NA>
#> 4        AlreadyDo                      3                <NA>
#> 5       VeryLikely                      1          VeryLikely
#> 6        AlreadyDo                      3                <NA>
#>   Q11_OtherChanges_en_dummy Q11_OtherChanges_Spec_en
#> 1                        NA                     <NA>
#> 2                        NA                     <NA>
#> 3                        NA                     <NA>
#> 4                        NA                     <NA>
#> 5                         1                     <NA>
#> 6                        NA                     <NA>
#>        Q12_MotivateChange_Category_en Q12_OtherChangeMotivation_Spec_en
#> 1 Diagnose;PersonalAdvice;WorseMemory                              <NA>
#> 2       Diagnose;Family;FamilySupport                              <NA>
#> 3    Family;FamilySupport;WorseMemory                              <NA>
#> 4      Family;ImpactKnown;WorseMemory                              <NA>
#> 5                Diagnose;WorseMemory                              <NA>
#> 6  Affordable;ImpactKnown;WorseMemory                              <NA>
#>     Q13_NoChange_Motivation_en Q13_SpecifyNoMotivation_Spec_en
#> 1   GiveupHobby;NotHelp;NoTime                            <NA>
#> 2 Alone;GiveupHobby;StartNoFun                            <NA>
#> 3  NoInfo;NoMotivation;NotHelp                            <NA>
#> 4               NotHelp;NoTime                            <NA>
#> 5                       NoInfo                            <NA>
#> 6     Expensive;NoInfo;NotHelp                            <NA>
#>    Q14_ChangeAssistance_en Q14_OtherAssistance_Spec_en Q15_FreetestAuth_en
#> 1 Checkup;GP;SupportFamily                        <NA>             Unknown
#> 2     Checkup;GP;MobileApp                        <NA>                 Yes
#> 3          GP;GroupSupport                        <NA>                 Yes
#> 4 Checkup;GP;SupportFamily                        <NA>                  No
#> 5 Checkup;GP;SupportFamily                        <NA>                 Yes
#> 6    Checkup;SupportFamily                        <NA>                 Yes
#>   Q15_FreetestAuth_en_dummy Q15_InformAuth_en Q15_InformAuth_en_dummy
#> 1                         3               Yes                       1
#> 2                         1               Yes                       1
#> 3                         1               Yes                       1
#> 4                         2               Yes                       1
#> 5                         1               Yes                       1
#> 6                         1               Yes                       1
#>   Q15_InsAuth_en Q15_InsAuth_en_dummy Q15_RelaxAuth_en Q15_RelaxAuth_en_dummy
#> 1            Yes                    1              Yes                      1
#> 2            Yes                    1              Yes                      1
#> 3             No                    2              Yes                      1
#> 4        Unknown                    3              Yes                      1
#> 5            Yes                    1              Yes                      1
#> 6            Yes                    1              Yes                      1
#>   Q15_SubsAuth_en Q15_SubsAuth_en_dummy Q15_TaxAuth_en Q15_TaxAuth_en_dummy
#> 1             Yes                     1            Yes                    1
#> 2             Yes                     1            Yes                    1
#> 3             Yes                     1            Yes                    1
#> 4             Yes                     1        Unknown                    3
#> 5         Unknown                     3             No                    2
#> 6             Yes                     1            Yes                    1
#>   Q16_BrainHealthSpec_en Q16_BrainHealthSpec_en_dummy   Q16_GP_en
#> 1            GreatExtent                            1 GreatExtent
#> 2            GreatExtent                            1 GreatExtent
#> 3            GreatExtent                            1     Neutral
#> 4            GreatExtent                            1     Neutral
#> 5                Neutral                            2 GreatExtent
#> 6            GreatExtent                            1     Neutral
#>   Q16_GP_en_dummy Q16_HealthAuthWebsites_en Q16_HealthAuthWebsites_en_dummy
#> 1               1               GreatExtent                               1
#> 2               1               GreatExtent                               1
#> 3               2               GreatExtent                               1
#> 4               2                   Neutral                               2
#> 5               1               GreatExtent                               1
#> 6               2                   Neutral                               2
#>   Q16_Newspaper_en Q16_Newspaper_en_dummy Q16_OnlineForum_en
#> 1          Neutral                      2            Neutral
#> 2      GreatExtent                      1           NotatAll
#> 3         NotatAll                      3            Neutral
#> 4         NotatAll                      3            Neutral
#> 5          Neutral                      2        GreatExtent
#> 6          Neutral                      2            Neutral
#>   Q16_OnlineForum_en_dummy Q16_PublicResearch_en Q16_PublicResearch_en_dummy
#> 1                        2               Neutral                           2
#> 2                        3           GreatExtent                           1
#> 3                        2               Neutral                           2
#> 4                        2               Neutral                           2
#> 5                        1               Neutral                           2
#> 6                        2           GreatExtent                           1
#>   Q16_ScientificJournals_en Q16_ScientificJournals_en_dummy Q16_SocialMedia_en
#> 1               GreatExtent                               1           NotatAll
#> 2               GreatExtent                               1            Neutral
#> 3               GreatExtent                               1            Neutral
#> 4                   Neutral                               2           NotatAll
#> 5                   Neutral                               2           NotatAll
#> 6               GreatExtent                               1           NotatAll
#>   Q16_SocialMedia_en_dummy Q16_TVRadio_en Q16_TVRadio_en_dummy
#> 1                        3        Neutral                    2
#> 2                        2        Neutral                    2
#> 3                        2    GreatExtent                    1
#> 4                        3       NotatAll                    3
#> 5                        3        Neutral                    2
#> 6                        3        Neutral                    2
#>   Q17_BrainResearch_Part_en Q18_Age_en Q18_Age_en_dummy Q19_Residence_en
#> 1                        No      41-60                3         Slovenia
#> 2                       Yes      41-60                3              USA
#> 3                        No      26-40                2         Slovenia
#> 4                       Yes      26-40                2          Denmark
#> 5                        No      26-40                2         Slovenia
#> 6                        No      41-60                3         Slovenia
#>   Q19_Residence_en_dummy Q20_Sex_en Q20_Sex_en_dummy Q21_Education_Coded_en
#> 1                    153       Male                2             GradSchool
#> 2                    179     Female                1             GradSchool
#> 3                    153       Male                2             GradSchool
#> 4                     45     Female                1             GradSchool
#> 5                    153       Male                2             GradSchool
#> 6                    153       Male                2             GradSchool
#>   Q21_Education_Coded_en_dummy Q22_Relationship_Status_en
#> 1                            5            DomesticPartner
#> 2                            5            DomesticPartner
#> 3                            5                     Single
#> 4                            5            DomesticPartner
#> 5                            5                  Unmarried
#> 6                            5            DomesticPartner
#>   Q22_Relationship_Status_en_dummy Q23_Vocation_categories_en
#> 1                                3                   Employed
#> 2                                3               Stay-at-home
#> 3                                1                   Employed
#> 4                                3                   Employed
#> 5                                2                   Employed
#> 6                                3                   Employed
#>   Q24_Exp_HealthCare_en Q25_Cognitive_Health_en Q26_Emotional_Health_en
#> 1                   Yes                 Average                 Average
#> 2                    No            AboveAverage            AboveAverage
#> 3                    No                 Average                 Average
#> 4                   Yes            AboveAverage               Excellent
#> 5                   Yes                 Average                 Average
#> 6                    No                 Average                 Average
#>   Q27_Illness_en Q28_BrainDisease_Care_en
#> 1             No                      Yes
#> 2            Yes                      Yes
#> 3             No                      Yes
#> 4             No                      Yes
#> 5             No                       No
#> 6             No                       No